updated dependencies; added tox config; refactored package name;
