+++ pyproject.toml
@@ -1,0 +1,130 @@
+[tool.poetry]
+name = "solvis-api"
+version = "0.1.0"
+description = "REST API for analysis of opensha modular Inversion Solutions"
+authors = ["Chris Chamberlain <chrisbc@artisan.co.nz>"]
+license = "AGPL"
+readme = "README.md"
+classifiers=[
+    'Development Status :: 2 - Pre-Alpha',
+    'Intended Audience :: Developers',
+    'License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
+    'Natural Language :: English',
+    'Programming Language :: Python :: 3',
+    'Programming Language :: Python :: 3.8',
+    'Programming Language :: Python :: 3.9',
+    'Programming Language :: Python :: 3.10',
+]
+packages = [{include = "solvis_api"}]
+
+[tool.poetry.dependencies]
+python = ">=3.8,<4.0"
+flask = "^2.2.2"
+Flask-Cors = "3.0.10"
+flask-restx = "^1.0.3"
+# geopandas = "0.10.2"
+# gql ="2.0.0"
+# nshm-toshi-client = "^0.0.0"
+fiona = "^1.8.22"
+nshm-toshi-client = "^1.0.0"
+pandas = "^1.5.2"
+geopandas = "^0.12.2"
+boto3 = "^1.26.33"
+solvis-store = {path = "/GNSDATA/API/solvis-store"}
+solvis = {path = "/GNSDATA/API/solvis"}
+
+
+[tool.poetry.group.dev.dependencies]
+tox = "^4.0.15"
+#boto3 = "^1.20.28"
+#botocore = "^1.23.28"
+moto = "^3.0.2"
+#nose = "^1.3.7"
+
+black  = { version = "^22.3"}
+isort  = { version = "^5.8.0"}
+flake8  = { version = "^3.9.2"}
+flake8-docstrings = { version = "^1.6.0", optional = true }
+mypy = {version = "^0.900"}
+pytest = { version = "^6.2.4"}
+pytest-cov  = { version = "^2.12.0"}
+# tox = { version = "^3.20.1"}
+virtualenv  = { version = "^20.2.2", optional = true}
+pip  = { version = "^20.3.1", optional = true}
+
+mkdocs  = { version = "^1.1.2"}
+mkdocs-include-markdown-plugin  = { version = "^1.0.0"}
+mkdocs-material  = { version = "^6.1.7"}
+mkdocstrings  = { version = "^0.18.0"}
+mkdocs-material-extensions  = { version = "^1.0.1"}
+mkdocs-autorefs = {version = "^0.3.1"}
+
+pymdown-extensions = {version = "^9.4"}
+mkdocstrings-python = {version = "^0.6.5"}
+mkdocs-pymdownx-material-extras = "^2.0.3"
+
+twine  = { version = "^3.3.0"}
+
+# pre-commit = {version = "^2.12.0", optional = true}
+# toml = {version = "^0.10.2", optional = true}
+# bump2version = {version = "^1.0.1"}
+# jinja2 = {version = "3.0.3"}
+# moto = "^3.1.10"
+# types-python-dateutil = "^2.8.16"
+
+[tool.poetry.extras]
+test = [
+    "pytest",
+    "black",
+    "isort",
+    "mypy",
+    "flake8",
+    "flake8-docstrings",
+    "pytest-cov"
+    ]
+
+dev = ["tox", "pre-commit", "virtualenv", "pip", "twine", "toml", "bump2version"]
+
+doc = [
+    "mkdocs",
+    "mkdocs-include-markdown-plugin",
+    "mkdocs-material",
+    "mkdocstrings",
+    "mkdocs-material-extension",
+    "mkdocs-autorefs"
+    ]
+
+[tool.black]
+line-length = 120
+skip-string-normalization = true
+target-version = ['py38', 'py39', 'py310']
+include = '\.pyi?$'
+exclude = '''
+/(
+    \.eggs
+  | \.git
+  | \.hg
+  | \.mypy_cache
+  | \.tox
+  | \.venv
+  | _build
+  | buck-out
+  | build
+  | dist
+)/
+'''
+
+[tool.isort]
+multi_line_output = 3
+include_trailing_comma = true
+force_grid_wrap = 0
+use_parentheses = true
+ensure_newline_before_comments = true
+line_length = 120
+skip_gitignore = true
+# you can skip files as below
+#skip_glob = docs/conf.py
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
