+++ solvis_api/logging.yaml
@@ -1,0 +1,53 @@
+version: 1
+disable_existing_loggers: False
+formatters:
+  simple:
+    format: "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
+
+handlers:
+  console:
+    class: logging.StreamHandler
+    level: DEBUG
+    formatter: simple
+    stream: ext://sys.stdout
+
+loggers:
+
+  flask:
+    level: WARNING
+    handlers: [console]
+    propagate: no
+
+  werkzeug:
+    level: INFO
+    handlers: [console]
+    propagate: no
+
+  api:
+    level: INFO
+    handlers: [console]
+    propagate: no
+
+  solvis_api.datastore:
+    level: WARNING
+    handlers: [console]
+    propagate: no
+
+  pynamodb:
+    level: INFO
+    handlers: [console]
+    propagate: no
+
+  process_solution:
+    level: DEBUG
+    handlers: [console]
+    propagate: no
+
+  aws_util:
+    level: DEBUG
+    handlers: [console]
+    propagate: no
+
+root:
+  level: INFO
+  handlers: [console]
